better detection of blacklisted errors
Exception handling returns a function name as opposed to the full
node with line numbers etc. This only seems to return on backtick
execution. So we record the function name that has triggered the error.

Then we search for that function name. It isn't perfect, but is going
to work most of the time.